+++ prototype-starter/app/filters.js
@@ -1,2 +1,10 @@
+//
+// For guidance on how to create filters see:
+// https://prototype-kit.service.gov.uk/docs/filters
+//
+
 const govukPrototypeKit = require('govuk-prototype-kit')
 const addFilter = govukPrototypeKit.views.addFilter
+
+// Add your filters here
+